Works but need to get the tracking better

+++ include/neat/genome.hpp
@@ -35,7 +35,7 @@ template <typename T> T &random_element(vector<T> &elems) {
     throw std::invalid_argument("No Elements");
   else if (elems.size() == 1)
     return elems[0];
-  int index = rand() % (elems.size() - 1);
+  size_t index = rand() % (elems.size() - 1);
   return elems[index];
 }
